requirements.txt
amqp==2.6.0
aniso8601==8.0.0
asgiref==3.2.10
attrs==19.3.0
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
clickclick==1.2.2
cligj==0.5.0
connexion==2.7.0
dataclasses==0.7
Django==3.0.5
django-cors-headers==3.4.0
djangorestframework==3.11.1
djangorestframework-stubs==1.2.0
djongo==1.3.3
Fiona==1.8.13
Flask==1.1.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
flask-restful-swagger==0.20.2
flask-restplus==0.13.0
freeze==3.0
future==0.18.2
GDAL==3.1.2
geopandas==0.8.1
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
kombu==4.6.11
MarkupSafe==1.1.1
marshmallow==3.7.1
mongoengine==0.20.0
munch==2.5.0
numpy==1.19.1
openapi-spec-validator==0.2.8
pandas==1.1.0
psycopg2==2.8.5
pymesh==1.0.2
pymongo==3.10.1
pyproj==2.6.1.post1
pyrsistent==0.16.0
pyshp==2.1.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
Shapely==1.7.0
six==1.15.0
sqlparse==0.2.4
swagger-ui==0.1.2
swagger-ui-bundle==0.0.8
urllib3==1.25.10
vine==1.3.0
Werkzeug==1.0.1
xlrd==1.2.0
yagmail==0.11.224
zipp==3.1.0
